Bug fixes
Re enabled interactions with buttons, as well as the leaderboard command.
