+++ src/commands/stats.js
@@ -10,11 +10,11 @@ function jsonProfileCruncher(json, killsJson) {
     var currentTopWeapon = 0
     var currentTopKill = 0
     // if killsJson is null, let top weapon be -------
-    if (killsJson == null) { 
+    if (killsJson == null) {
         topWeaponString = '-------'
     } else {
         // finds the "id" of the weapon with the most "kills"
-        for(var i = 0; i < killsJson.length; i++) {
+        for (var i = 0; i < killsJson.length; i++) {
             if (parseInt(killsJson[i]['kills']) > currentTopKill) {
                 currentTopWeapon = killsJson[i]['id']
                 currentTopKill = killsJson[i]['kills']
@@ -57,22 +57,22 @@ function jsonProfileCruncher(json, killsJson) {
     }
     if (json['clan_id'] != null) { clan = `${json['clan_role']} of ${json['clan_name']}(${json['clan_id']})\n` }
     return `**Username:** \`${json['name']}\`\n` +
-            `**UserID:** ${json['id']}\n` +
-            `**Status:** ${json['status']}\n` +
-            `**Level:** ${functions.levelMaker(json['experience'])}\n` +
-            `**Exp:** ${nf.format(json['experience'])}\n` +
-            `**Level Rank:** ${nf.format(json['ranking'])}\n` +
-            `**Kills:** ${nf.format(json['kills'])}\n` +
-            `**Deaths:** ${nf.format(json['deaths'])}\n` +
-            `**K/D Ratio:** ${Math.round(1000 * (parseInt(json['kills']) / parseInt(json['deaths']))) / 1000 || 0}\n` +
-            `**Flag Captures:** ${nf.format(json['caps'])}\n` +
-            `**Title:** ${functions.mapToRankTitles(json['skill'])}\n` +
-            `**Skill Points:** ${nf.format(json['skill'])}\n` +
-            `**Skill Rank:** ${nf.format(json['skill_ranking'])}\n` +
-            `**Top Weapon:** ${topWeaponString}\n` +
-            `Created on <t:${(Date.parse(json['created']) / 1000) - 18000}:D>\n` +
-            `Last seen <t:${(Date.parse(json['seen']) / 1000) - 18000}:R>\n` +
-            `${clan}`
+        `**UserID:** ${json['id']}\n` +
+        `**Status:** ${json['status']}\n` +
+        `**Level:** ${functions.levelMaker(json['experience'])}\n` +
+        `**Exp:** ${nf.format(json['experience'])}\n` +
+        `**Level Rank:** ${nf.format(json['ranking'])}\n` +
+        `**Kills:** ${nf.format(json['kills'])}\n` +
+        `**Deaths:** ${nf.format(json['deaths'])}\n` +
+        `**K/D Ratio:** ${Math.round(1000 * (parseInt(json['kills']) / parseInt(json['deaths']))) / 1000 || 0}\n` +
+        `**Flag Captures:** ${nf.format(json['caps'])}\n` +
+        `**Title:** ${functions.mapToRankTitles(json['skill'])}\n` +
+        `**Skill Points:** ${nf.format(json['skill'])}\n` +
+        `**Skill Rank:** ${nf.format(json['skill_ranking'])}\n` +
+        `**Top Weapon:** ${topWeaponString}\n` +
+        `Created on <t:${(Date.parse(json['created']) / 1000) - 18000}:D>\n` +
+        `Last seen <t:${(Date.parse(json['seen']) / 1000) - 18000}:R>\n` +
+        `${clan}`
 }
 
 
@@ -85,33 +85,33 @@ function weaponStatsMakeup(json, name) {
     }
     var title = `\`${name}\` has gotten the following kills:\n`
     var send = `**Fists:** ${nf.format(weaponDict['-1'])}\n` +
-                `**Shotgun:** ${nf.format(weaponDict['1'])}\n` +
-                `**SMG:** ${nf.format(weaponDict['2'])}\n` +
-                `**M79:** ${nf.format(weaponDict['3'])}\n` +
-                `**Barrett:** ${nf.format(weaponDict['4'])}\n` +
-                `**Shock Rifle:** ${nf.format(weaponDict['5'])}\n` +
-                `**Pulse Gun:** ${nf.format(weaponDict['6'])}\n` +
-                `**Flamer:** ${nf.format(weaponDict['7'])}\n` +
-                `**RPG:** ${nf.format(weaponDict['8'])}\n` +
-                `**Rifle:** ${nf.format(weaponDict['9'])}\n` +
-                `**Lasergun:** ${nf.format(weaponDict['10'])}\n` +
-                `**AK-47:** ${nf.format(weaponDict['12'])}\n` +
-                `**Hand Grenade:** ${nf.format(weaponDict['20'])}\n` +
-                `**Cluster Grenade:** ${nf.format(weaponDict['21'])}\n` +
-                `**Shuriken:** ${nf.format(weaponDict['23'])}\n` +
-                `**Deagles:** ${nf.format(weaponDict['24'])}\n` +
-                `**Snowballs:** ${nf.format(weaponDict['25'])}\n` +
-                `**Minigun:** ${nf.format(weaponDict['26'])}\n` +
-                `**X75:** ${nf.format(weaponDict['27'])}\n` +
-                `**MAC-10:** ${nf.format(weaponDict['28'])}\n` +
-                `**Bow:** ${nf.format(weaponDict['29'])}\n` +
-                `**Avenger:** ${nf.format(weaponDict['30'])}\n` +
-                `**Carbine:** ${nf.format(weaponDict['31'])}\n` +
-                `**Chainsaw:** ${nf.format(weaponDict['204'])}\n` +
-                `**Link Gun:** ${nf.format(weaponDict['213'])}\n` +
-                `**Boomerang:** ${nf.format(weaponDict['33'])}\n` +
-                `**Uzi:** ${nf.format(weaponDict['35'])}\n` +
-                `**M60:** ${nf.format(weaponDict['34'])}`
+        `**Shotgun:** ${nf.format(weaponDict['1'])}\n` +
+        `**SMG:** ${nf.format(weaponDict['2'])}\n` +
+        `**M79:** ${nf.format(weaponDict['3'])}\n` +
+        `**Barrett:** ${nf.format(weaponDict['4'])}\n` +
+        `**Shock Rifle:** ${nf.format(weaponDict['5'])}\n` +
+        `**Pulse Gun:** ${nf.format(weaponDict['6'])}\n` +
+        `**Flamer:** ${nf.format(weaponDict['7'])}\n` +
+        `**RPG:** ${nf.format(weaponDict['8'])}\n` +
+        `**Rifle:** ${nf.format(weaponDict['9'])}\n` +
+        `**Lasergun:** ${nf.format(weaponDict['10'])}\n` +
+        `**AK-47:** ${nf.format(weaponDict['12'])}\n` +
+        `**Hand Grenade:** ${nf.format(weaponDict['20'])}\n` +
+        `**Cluster Grenade:** ${nf.format(weaponDict['21'])}\n` +
+        `**Shuriken:** ${nf.format(weaponDict['23'])}\n` +
+        `**Deagles:** ${nf.format(weaponDict['24'])}\n` +
+        `**Snowballs:** ${nf.format(weaponDict['25'])}\n` +
+        `**Minigun:** ${nf.format(weaponDict['26'])}\n` +
+        `**X75:** ${nf.format(weaponDict['27'])}\n` +
+        `**MAC-10:** ${nf.format(weaponDict['28'])}\n` +
+        `**Bow:** ${nf.format(weaponDict['29'])}\n` +
+        `**Avenger:** ${nf.format(weaponDict['30'])}\n` +
+        `**Carbine:** ${nf.format(weaponDict['31'])}\n` +
+        `**Chainsaw:** ${nf.format(weaponDict['204'])}\n` +
+        `**Link Gun:** ${nf.format(weaponDict['213'])}\n` +
+        `**Boomerang:** ${nf.format(weaponDict['33'])}\n` +
+        `**Uzi:** ${nf.format(weaponDict['35'])}\n` +
+        `**M60:** ${nf.format(weaponDict['34'])}`
 
     send = title + send;
 
@@ -155,12 +155,12 @@ module.exports = {
             pingEm = new EmbedBuilder()
                 .setColor(interaction.guild.members.me.displayHexColor)
                 .setDescription('Something went critically wrong. Please try again later.')
-            await interaction.editReply({ embeds: [pingEm]});
+            await interaction.editReply({ embeds: [pingEm] });
         } else if (profile == 'badName') {
             pingEm = new EmbedBuilder()
                 .setColor(interaction.guild.members.me.displayHexColor)
                 .setDescription('I could not find a player with that name, please double check your spelling.\nIf you unsure of the spelling, you can try searching for them using </search:1051343380069752909>.')
-            await interaction.editReply({ embeds: [pingEm]});
+            await interaction.editReply({ embeds: [pingEm] });
         } else {
             pingEm = new EmbedBuilder()
                 .setColor(color)
@@ -172,8 +172,7 @@ module.exports = {
                         .setStyle('Secondary')
                         .setLabel(`View ${profile['name']}'s weapon statistics`)
                         .setDisabled(false))
-            // FIXME: await interaction.editReply({ embeds: [pingEm], components: [row]})
-            await interaction.editReply({ embeds: [pingEm] })
+            await interaction.editReply({ embeds: [pingEm], components: [row] })
 
             // pause for a tenth of a second
             await new Promise(r => setTimeout(r, 100));
@@ -183,12 +182,11 @@ module.exports = {
             EmWithTopWeapon = new EmbedBuilder()
                 .setColor(color)
                 .setDescription(jsonProfileCruncher(profile, weaponStats))
-            // FIXME: await interaction.editReply({ embeds: [EmWithTopWeapon], components: [row]})
-            await interaction.editReply({ embeds: [EmWithTopWeapon] })
-            
+            await interaction.editReply({ embeds: [EmWithTopWeapon], components: [row] })
+
         }
 
-        
+
 
 
         const filter = i => i.customId === 'expand' && i.user.id == interaction.user.id;
